+++ lib/src/client_types/client_type.dart
@@ -1,0 +1,7 @@
+/// A Client maybe for personal use (only a single person can log in),
+/// or for public use.
+///
+/// As of right now, public clients are not available and the class below is
+/// therefore commented out. This is to be added back in future when the
+/// implementation comes out.
+class ClientType {}