+++ lib/src/constants/authority.dart
@@ -1,1 +1,2 @@
 const AUTHORITY = 'api.mangadex.org';
+const AUTH_AUTHORITY = 'auth.mangadex.org';